@@ -1,31 +1,31 @@
 package com.polly.myapplication
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.ViewGroup
 
-class CustomLayoutManager4(val screenWidth: Int, val viewWidth: Int) : RecyclerView.LayoutManager() {
+class CustomLayoutManager4(val screenWidth: Int, val viewWidth: Int) : androidx.recyclerview.widget.RecyclerView.LayoutManager() {
 
     private var horizontalScrollOffset: Int = 0
 
-    override fun generateDefaultLayoutParams(): RecyclerView.LayoutParams {
-        return RecyclerView.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT)
+    override fun generateDefaultLayoutParams(): androidx.recyclerview.widget.RecyclerView.LayoutParams {
+        return androidx.recyclerview.widget.RecyclerView.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT)
     }
 
     override fun canScrollHorizontally(): Boolean {
         return true
     }
 
-    override fun scrollHorizontallyBy(dx: Int, recycler: RecyclerView.Recycler, state: RecyclerView.State): Int {
+    override fun scrollHorizontallyBy(dx: Int, recycler: androidx.recyclerview.widget.RecyclerView.Recycler, state: androidx.recyclerview.widget.RecyclerView.State): Int {
         horizontalScrollOffset += dx
         fill(recycler, state)
         return dx
     }
 
-    override fun onLayoutChildren(recycler: RecyclerView.Recycler, state: RecyclerView.State) {
+    override fun onLayoutChildren(recycler: androidx.recyclerview.widget.RecyclerView.Recycler, state: androidx.recyclerview.widget.RecyclerView.State) {
         fill(recycler, state)
     }
 
-    private fun fill(recycler: RecyclerView.Recycler, state: RecyclerView.State) {
+    private fun fill(recycler: androidx.recyclerview.widget.RecyclerView.Recycler, state: androidx.recyclerview.widget.RecyclerView.State) {
         detachAndScrapAttachedViews(recycler)
 
         var firstVisiblePosition = horizontalScrollOffset / viewWidth
